updated message.form.tsx to main.contend.tsx and included the Logo and continue button
